// fe/src/lib/api/generated/model/goodMoralCorrectionPayload.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * COMPASS API
 * The version-one backend API for COMPASS.
 * OpenAPI spec version: 1.0.0
 */

export interface GoodMoralCorrectionPayload {
  applicant_name?: string | null;
  college?: string | null;
  course?: string | null;
  degree?: string | null;
  graduation_date?: string | null;
  major?: string | null;
  official_receipt_amount?: number | string | null;
  official_receipt_date?: string | null;
  official_receipt_number?: string | null;
  semester?: string | null;
  year_level?: string | null;
}