// fe/src/lib/api/generated/model/goodMoralDetailResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * COMPASS API
 * The version-one backend API for COMPASS.
 * OpenAPI spec version: 1.0.0
 */
import type { AcademicYearSummary } from './academicYearSummary';
import type { FormRevisionSummary } from './formRevisionSummary';
import type { GoodMoralStatusValue } from './goodMoralStatusValue';
import type { GoodMoralVariantValue } from './goodMoralVariantValue';
import type { PersonSummary } from './personSummary';

export interface GoodMoralDetailResponse {
  academic_year: AcademicYearSummary | null;
  applicant_name: string;
  college: string;
  course: string;
  created_at: string;
  degree: string;
  document_template_key: string | null;
  document_template_version: number | null;
  form_revision: FormRevisionSummary | null;
  graduation_date: string | null;
  id: string;
  inventory_id: string | null;
  issued_at: string | null;
  issued_by: PersonSummary | null;
  issued_by_name_snapshot: string;
  major: string;
  official_receipt_amount: string | null;
  official_receipt_date: string | null;
  official_receipt_number: string;
  semester: string;
  status: GoodMoralStatusValue;
  student: PersonSummary;
  updated_at: string;
  variant: GoodMoralVariantValue;
  year_level: string;
}